fix(studio.giselles.ai): Fix build error
studio.giselles.ai:build: Error occurred prerendering page "/auth/github/complete". Read more: https://nextjs.org/docs/messages/prerender-error
studio.giselles.ai:build: Export encountered an error on /auth/github/complete/page: /auth/github/complete, exiting the build.
